Split Artifact to have ArtifactCreate that handles only the creation part, not the run part
