+++ lib/engine-dylib/Cargo.toml
@@ -11,6 +11,7 @@ readme = "README.md"
 edition = "2018"
 
 [dependencies]
+wasmer-artifact = { path = "../artifact", version = "=2.2.1" }
 wasmer-types = { path = "../types", version = "=2.2.1" }
 wasmer-compiler = { path = "../compiler", version = "=2.2.1" }
 wasmer-vm = { path = "../vm", version = "=2.2.1", features = ["enable-rkyv"] }
