+++ lib/engine/Cargo.toml
@@ -13,6 +13,7 @@ edition = "2018"
 [dependencies]
 wasmer-types = { path = "../types", version = "=2.2.1" }
 wasmer-compiler = { path = "../compiler", version = "=2.2.1" }
+wasmer-artifact = { path = "../artifact", version = "=2.2.1" }
 target-lexicon = { version = "0.12.2", default-features = false }
 # flexbuffers = { path = "../../../flatbuffers/rust/flexbuffers", version = "0.1.0" }
 backtrace = "0.3"
