+++ lib/vm/Cargo.toml
@@ -12,6 +12,7 @@ edition = "2018"
 
 [dependencies]
 wasmer-types = { path = "../types", version = "=2.2.1" }
+wasmer-artifact = { path = "../artifact", version = "=2.2.1" }
 libc = { version = "^0.2", default-features = false }
 memoffset = "0.6"
 indexmap = { version = "1.6", features = ["serde-1"] }
